backend/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
pydantic
python-magic
PyPDF2
python-docx
python-multipart
aiohttp
beautifulsoup4
brotli
trafilatura
readability-lxml
html2text
httpx[http2]
pydantic[dotenv]
python-dotenv
